return books page is done and only text mode in form page left

+++ database/migrations/2018_09_15_113952_create_students_table.php
@@ -17,7 +17,7 @@ class CreateStudentsTable extends Migration
             $table->increments('id');
            $table->integer('students_class_id');
             $table->integer('students_name_id');
-            $table->integer('contact_number');
+            $table->bigInteger('contact_number');
             $table->timestamps();
         });
     }
